# server/swagger.yaml
swagger: '2.0'
info:
  description: This is a simple API To CRUS Jobs fetched from upwork RSS
  version: 1.0.0
  title: Gigs For Geeks API

  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html

paths:
  /jobs:
    get:
      summary: Fetch latest jobs from upwork
      operationId: getJobs
      description: |
        By passing in the appropriate options, you can fetch upwork jobs fetched by the system
      produces:
      - application/json
      parameters:
      - in: query
        name: page
        description: Pass an optional page index for pagination
        required: false
        default: 1
        type: integer
      - name: authorization
        in: header
        description: An authorization header holding bearerToken
        required: true
        type: string
      responses:
        200:
          description: upwork jobs fetched by the system
          schema:
            type: object
            properties:
             page:
              type: integer
              example: 1
             total:
              type: integer
              example: 220
             results:
              type: array
              items:
               $ref: '#/definitions/JobItem'
             cursor:
              type: string
              example: https://host:port/api/jobs?page=2
        400:
          description: bad input parameter
          schema:
            type: object
            properties:
             success:
              type: integer
              example: false
             message:
              type: string
        401:
          description: unauthorized
          schema:
            type: object
            properties:
             success:
              type: integer
              example: false
             message:
              type: string
              example: token expired
  /jobs/{id}:
    delete:
      summary: Delete a job given it's id
      operationId: deleteJob
      description: Delete a job from system saved jobs
      parameters:
      - in: path
        required: true
        type: number
        name: id
        description: Job to be deleted
      - name: authorization
        in: header
        description: An authorization header holding bearerToken
        required: true
        type: string
      responses:
        204:
          description: job deleted
        400:
          description: bad input parameter
          schema:
            type: object
            properties:
             success:
              type: integer
              example: false
             message:
              type: string
        401:
          description: unauthorized
          schema:
            type: object
            properties:
             success:
              type: integer
              example: false
             message:
              type: string
              example: token expired
definitions:
  JobItem:
    type: object
    required:
    - id
    - title
    - link
    - content
    - isoDate
    properties:
      id:
        type: number
        example: 120
      title:
        type: string
        example: Full stack developer for an on going project
      link:
        type: string
        example: https://www.upwork.com/jobs/Full-Stack-Node-preffered-dev-needed_%7E01afd235e4dfc3fc60?source=rss
      content:
        type: string
        example: "Hello!<br /> Our small winery is looking for a developer to build a classy website. Basically we need a service where our clients can buy our products, rate them and compare the prices with other wineries. Also we&rsquo;d like to have the ability to send a newsletter to those who signed up for it. Please share your most relevant projects and start with the word &ldquo;crocodile&rdquo;.<br /><br /><br /><b>Posted On</b>: June 05, 2020 14:04 UTC<br /><b>Category</b>: Full Stack Development<br /><b>Skills</b>:eCommerce, Database, JavaScript, PHP, MySQL, Microsoft SQL Server, RESTful API, MongoDB, API <br /><b>Country</b>: Austria <br /><a >click to apply</a>"
      isoDate:
        type: string
        example: "016-08-29T09:12:33.001Z"

# Added by API Auto Mocking Plugin
host: virtserver.swaggerhub.com
basePath: /Yousif-Hmada/gig_for_geeks/1.0.0
schemes:
 - https